SHT3X driver: change state timer variable from uint32 to uint64.
Prevents current behaviour of repeated entry into 'init' state
and associated log spamming after +-72 minutes.

+++ src/drivers/hygrometer/sht3x/sht3x.h
@@ -110,7 +110,7 @@ private:
 
 	float measured_temperature = 0;
 	float measured_humidity = 0;
-	uint32_t measurement_time = 0;
+	uint64_t measurement_time = 0;
 	uint16_t measurement_index = 0;
 
 	sht_info _sht_info;
